+++ libc/ipc/precv.c
@@ -29,7 +29,7 @@
  */
 
 static struct packet *_precvm
-	(uint8_t port, uint32_t source, uint64_t inode, uint16_t id, uint16_t frag) {
+	(uint8_t port, uint32_t source, uint32_t inode, uint16_t id, uint16_t frag) {
 	struct message *m;
 	struct packet *packet;
 	bool match = false;
@@ -104,7 +104,7 @@ struct packet *precvs(uint8_t port, uint32_t source) {
  * precvn
  */
 
-struct packet *precvn(uint8_t port, uint32_t source, uint64_t inode) {
+struct packet *precvn(uint8_t port, uint32_t source, uint32_t inode) {
 	return _precvm(port, source, inode, 0, -1);
 }
